pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["jgabriel1 <josegabriel@det.ufc.br>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["python-multipart"], version = "^0.56.0"}
uvicorn = "^0.11.5"
requests = "^2.22.0"
pymongo = "^3.10.1"
pyjwt = "^1.7.1"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
dnspython = "^1.16.0"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.5.1"}
aiohttp = "^3.6.2"
python-dotenv = "^0.14.0"
motor = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4"
flake8 = "^3.8.3"
autopep8 = "^1.5.3"

[tool.poetry.scripts]
start = "scripts:start"
dev = "scripts:dev"
test = "scripts:test"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
